Rename keymgmt-key to keymgmt/key
